refactor: extract PersonaFields into shared component
